Update and rename IntertekForm/IntertekForm/Page1VC.swift to IntertekForm/UserInfoViewController.swift
MVC refactor.
